### 2025-01-09 #172 @editorialbot
[PRE REVIEW]: ExponentialFamilyManifolds.jl: Representing exponential families as Riemannian manifolds
**Submitting author:** @Nimrais (<a href="http://orcid.org/0009-0003-3011-0825">Mykola Lukashchuk</a>)
**Repository:** https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.5.0
**Managing EiC:** Luca Ferranti

Thanks for submitting your paper to JCON @Nimrais. **Currently, there isn't a JCON editor assigned** to your paper.
