### 2025-01-15 #172 @lucaferranti
Hi @Nimrais :wave: ,

I will be editor for your JuliaCon proceedings submission. @kellertuer has already kindly volunteered to review this. Once I find a second reviewer, I'll create the review issues where reviewers can leave comments, discuss, etc.

If you have any questions, let me know
