### 2025-01-15 #172 @Nimrais
Hi @lucaferranti, Thanks! I am really happy that [kellertuer](https://github.com/kellertuer) volunteered to review the package. I suggest having some people with statistical backgrounds review the package as well. The following people look like good options: xieyj17, gvdr, ayushpatnaikgit, and sourish-cmi.
