### 2025-01-20 #172 @lucaferranti
Hi @ayushpatnaikgit 👋 ,

would you like to review this paper for JuliaCon proceedings?

Reviewing for JuliaCon proceedings is simila to reviewing for JOSS (Journal of Open Source Software). In addition to reading the paper and commenting on it, it involves playing with the code, commenting on repository quality and similar. People tend to consider it somehow more enjoyable than traditional academic peer review. I will also be available through the whole process to answer any question you may have. Ideally, we encourage reviewers to get the review completed within a few week (3-4).

let me know if you have further questions
