### 2025-01-22 #179 @editorialbot
[REVIEW]: ExponentialFamilyManifolds.jl: Representing exponential families as Riemannian manifolds
**Submitting author:** @Nimrais (<a href="http://orcid.org/0009-0003-3011-0825">Mykola Lukashchuk</a>)
**Repository:** https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v3.0.1
