### 2025-01-22 #179 @lucaferranti
Hi  @Nimrais 👋 and thank you for submitting to JuliaCon proceedings.

I will be the editor handling this submission and this issue is for the review.

Huge thanks to @kellertuer and @olivierverdier for volunteering as reviewers 🙏 . To get started you can generate your reviewer checklist by commenting

on this issue.

As you go through the checklist points, you can leave your comments for the author to address either in this issue or as issues on the software repository. If you do the latter, please make sure to link to this issue too.

If you have any questions at any point, do not hesitate to ping me and ask.
