### 2025-01-22 3576211
fix: implement base normal gamma manifold
