+++ test/normal_gamma_tests.jl
@@ -1,0 +1,156 @@
+@testitem "Partial derivatives of NormalGamma Fisher metric in natural coords" begin
+
+    import ExponentialFamilyManifolds: partial_metric
+    using ExponentialFamily
+    using ForwardDiff
+    using BenchmarkTools
+
+    # define some test points in domain:
+    test_points = [
+        [1.0, -2.0, 0.5, -1.0],
+        [1.5, -1.0, 0.3, -0.9],
+        [2.0, -3.0, 1.0, -2.0],
+        [0.5, -0.5, 0.1, -0.1],
+    ]
+
+    d = NormalGamma(1.0, 1.0, 1.0, 1.0)
+    ef = convert(ExponentialFamilyDistribution, d)
+    locfisher = (η) -> fisherinformation(ef, η)
+
+
+     # A helper that flattens
+    function fisher_metric_vec(η)
+        G = locfisher(η)
+        return vec(G)
+    end
+
+    function forwarddiff_partials(η)
+        FD_jac = ForwardDiff.jacobian(fisher_metric_vec, η)
+        out = Array{Float64}(undef, 4,4,4)
+        for i in 1:4
+            for j in 1:4
+                r = 4*(i-1)+j
+                for c in 1:4
+                    out[i,j,c] = FD_jac[r,c]
+                end
+            end
+        end
+        return out
+    end
+
+    for η in test_points
+        fd_partials = forwarddiff_partials(η)
+        sym_partials = partial_metric(NormalGamma, NaturalParametersSpace(), η)
+
+        diff_array = fd_partials .- sym_partials
+        err = maximum(abs, diff_array)
+
+        @test err < 1e-6  # or smaller threshold if you want more precision
+
+        fd_bench = @benchmark forwarddiff_partials($η)
+        sym_bench = @benchmark partial_metric(NormalGamma, NaturalParametersSpace(), $η)
+        @test min(sym_bench.times...) < min(fd_bench.times...)
+        #TODO WHY?
+        # probably smt just wrong here
+        @test_broken sym_bench.allocs < fd_bench.allocs
+    end
+end
+
+@testitem "Generic properties of NormalGammaNaturalManifold" begin
+    import ExponentialFamilyManifolds: NormalGammaNaturalManifold
+    import ManifoldsBase: manifold_dimension, representation_size, check_point, check_vector, is_point, is_vector
+    import ManifoldsBase: zero_vector, inner
+    import ExponentialFamilyManifolds: isproper, NaturalParametersSpace
+
+    using ManifoldsBase, Test
+    using LinearAlgebra
+    using ExponentialFamily
+
+    M = NormalGammaNaturalManifold()
+    valid_point = [1.0, -2.0, 0.5, -1.0]
+    ef = ExponentialFamilyDistribution(NormalGamma, [1.0, -2.0, 0.5, -1.0], nothing)
+
+    @test @inferred(manifold_dimension(M)) === 4
+    @test @inferred(representation_size(M)) === (4,)
+    @test injectivity_radius(M) === Inf
+
+    @test check_point(M, valid_point) === nothing
+    @test check_point(M, [2.0, 0.0, 1.0, -3.0]) isa DomainError # invalid because η2>=0
+    @test check_point(M, [0.5, -3.0, -1.0, -1.0]) isa DomainError # invalid because η3 <= -1/2
+    @test check_point(M, [1, 1, 1]) isa DomainError # invalid because because length(η) != 4
[…]
