### 2025-01-22 d20a64e
fix: implement normal gamma natural manifold
