### 2025-01-22 786593e
fix: remove NLsolve
