### 2025-01-22 9a89941
fix: add benchmark tools into tests
