### 2025-01-22 1519f72
fix: add rand operation for normal gamma manifold
