### 2025-01-22 9e80b86
fix: use manifolds test set
