### 2025-01-26 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @kellertuer
In a second part I will add a few comments on the documentation and the code in a few days.
