### 2025-01-27 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @Nimrais
@olivierverdier Thank you for the review. Before I address other parts of your review, there is some misunderstanding about what the package already implements. The paper directly provides examples of two different multivariate normal families.  Also, you can find that they are indeed implemented in the code https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl/blob/main/src/natural_manifolds/normal.jl#L27-L32 and https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl/blob/main/src/natural_manifolds/normal.jl#L51-L56.
