### 2025-01-27 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @olivierverdier
Fair enough, sorry I overlooked those!
