### 2025-02-06 dcb86f8
CompatHelper: bump compat for ManifoldsBase to 1, (keep existing compat)
