+++ Project.toml
@@ -18,7 +18,7 @@ BayesBase = "1.3"
 ExponentialFamily = "1.6.0"
 LinearAlgebra = "1.10"
 Manifolds = "0.9"
-ManifoldsBase = "0.15"
+ManifoldsBase = "0.15, 1"
 Random = "1.10"
 RecursiveArrayTools = "3"
 Static = "0.8, 1"
