### 2025-02-07 13d0d1c
fix: ensure positivity should quite big, some time drops NaN
