### 2025-02-07 45c87e7
refactor: re-use code from Manifolds.jl
