### 2025-02-07 45c1393
fix: normal gamma is an open set in R^{4}, check that projection is indentity
