### 2025-02-07 1cbf2a0
docs: add NormalGammaNaturalManifold into docs
