### 2025-02-07 86da7fd
docs: improve NormalGammaNaturalManifold docstring
