### 2025-02-07 f9beac2
docs: point to correct place for NormalGamma distribution
