### 2025-02-11 e0f5ebc
fix: update package according Manifolds updates
