+++ src/shifted_negative_numbers.jl
@@ -1,4 +1,3 @@
-
 """
     ShiftedNegativeNumbers(shift)
 
@@ -56,14 +55,18 @@ function ManifoldsBase.inner(M::ShiftedNegativeNumbers, p, X, Y)
     )
 end
 
-function ManifoldsBase.exp!(M::ShiftedNegativeNumbers, q, p, X, t::Number=1)
+function ManifoldsBase.exp_fused!(M::ShiftedNegativeNumbers, q, p, X, t::Number)
     @inbounds q[1] = shift(
         M,
-        -ManifoldsBase.exp(
+        -ManifoldsBase.exp_fused(
             PositiveNumbers(), -unshift(M, @inbounds(p[1])), -@inbounds(X[1]), t
         ),
     )
     return q
+end
+
+function ManifoldsBase.exp!(M::ShiftedNegativeNumbers, q, p, X)
+    return ManifoldsBase.exp_fused!(M, q, p, X, one(eltype(p)))
 end
 
 function ManifoldsBase.log!(M::ShiftedNegativeNumbers, X, p, q)
@@ -82,10 +85,14 @@ end
 
 ManifoldsBase.default_retraction_method(::ShiftedNegativeNumbers) = ExponentialRetraction()
 
-function ManifoldsBase.retract!(
+function ManifoldsBase.retract_fused!(
     M::ShiftedNegativeNumbers, q, p, X, t::Number, ::ExponentialRetraction
 )
-    return ManifoldsBase.exp!(M, q, p, X, t)
+    return ManifoldsBase.exp_fused!(M, q, p, X, t)
+end
+
+function ManifoldsBase.retract!(M::ShiftedNegativeNumbers, q, p, X, ::ExponentialRetraction)
+    return ManifoldsBase.retract_fused!(M, q, p, X, one(eltype(p)), ExponentialRetraction())
 end
 
 function ManifoldsBase.parallel_transport_to!(M::ShiftedNegativeNumbers, Y, p, X, q)
