### 2025-02-11 745c852
test: retract! ≈ retract_fused
