+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ExponentialFamilyManifolds"
 uuid = "5c9727c4-3b82-4ab3-b165-76e2eb971b08"
 authors = ["Mykola Lukashchuk <m.lukashchuk@tue.nl>", "Dmitry Bagaev <bvdmitri@gmail.com>", "Albert Podusenko <albert@lazydynamics.com>"]
-version = "1.5.1"
+version = "2.0.0"
 
 [deps]
 BayesBase = "b4ee3484-f114-42fe-b91c-797d54a0c67e"
@@ -15,7 +15,7 @@ Static = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
 
 [compat]
 BayesBase = "1.3"
-ExponentialFamily = "1.6.0"
+ExponentialFamily = "2.0.0"
 LinearAlgebra = "1.10"
 Manifolds = "0.10"
 ManifoldsBase = "1"
