### 2025-02-20 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @Nimrais
Thank you for this insightful review about manifold representation independence. However, I agree with your main point: `a manifold is independent of the representation`. However, from my point of view, the real manifold is a set of exponential distributions specified with base measure, sufficient statistics, and domain. So, however, as you correctly noted,` shifted negative numbers` makes no sense; representing them only through manifold from `Manifolds.jl` also makes little sense. It is a coincidence that some coordinates are the same, and I can reuse them.

There are two reasons for that: (1) I want to avoid the transformation of natural parameters into some manifolds-friendly form (for example, for multivariate normal, it becomes too slow I need to invert the matrix on each iteration, and (2) Fisher information matrix in natural parameters is just a hessian of the log partition, so to evaluate I do not need to integrate.

The key insight here is that while geometrically, these might be the same manifolds, the statistical structure (Fisher metric, exponential family structure) provides additional context that makes specific coordinate systems computationally privileged. Natural parameters aren't just another coordinate system - they're the coordinate system where the Fisher information takes its most straightforward form from the computational point of view.

So, to accommodate these two goals (don't change coordinates too much) and be able to compute fisher information only through derivation, my refactoring proposal for this issue is the following one:

```julia
# The manifold is the space of distributions
struct ExponentialFamilyManifold{𝔽,T,D,C} <: AbstractManifold{𝔽}
    dims::D
    conditioner::C
end

# Natural parameters form an atlas
struct NaturalParametersAtlas{T} <: AbstractAtlas{ℝ} end

# Fisher Information as the metric
struct FisherInformationMetric <: RiemannianMetric end

# Decorate with the Fisher metric
const StatisticalManifold{𝔽,T,D,C} = MetricManifold{𝔽,ExponentialFamilyManifold{𝔽,T,D,C},FisherInformationMetric}

# Implementation of chart operations
function Manifolds.get_parameters(M::ExponentialFamilyManifold, A::NaturalParametersAtlas, i, p)
    # Convert distribution p to its natural parameters
    return natural_parameters(p)
end

function Manifolds.get_point(M::ExponentialFamilyManifold, A::NaturalParametersAtlas, i, η)
    # Convert natural parameters η back to the distribution
    return distribution_from_natural(M, η)
end

# Implement the Fisher metric - it's straightforward in natural parameters
function Manifolds.inner(
    M::StatisticalManifold,
    p,
    X,
    Y,
    A::NaturalParametersAtlas = get_default_atlas(M)
)
    i = get_chart_index(M, A, p)
    η = get_parameters(M, A, i, p)
    # In natural parameters, Fisher metric is just the Hessian of log partition function
    return dot(X, hessian_log_partition(M, η) * Y)
end
```

This approach allows the implementation of another set of coordinates if needed and, as an extra bonus, allows the implementation of an approximation to an exponential map on the `StatisticalManifold:`

```julia
struct ChartFirstOrderRetraction{TA<:AbstractAtlas} <: AbstractRetractionMethod
    A::TA
end

# Second-order retraction in charts
struct ChartSecondOrderRetraction{TA<:AbstractAtlas} <: AbstractRetractionMethod
    A::TA
end

# Implementation of statistical manifold
function retract_chart_first_order!(M::StatisticalManifold, q, p, X, m::ChartFirstOrderRetraction)
    i = get_chart_index(M, m.A, p)
    p_chart = get_parameters(M, m.A, i, p)
    X_chart = get_coordinates(M, p, X, induced_basis(M, m.A, i))
    # Simple addition in chart coordinates
    q_chart = p_chart + X_chart
    return get_point!(M, q, m.A, i, q_chart)
end
```

For a rough draft of this implementation, you can check https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl/pull/26 for one specific distribution, Normal-Gamma.

So, would such an implementation be able to resolve your concern?
