### 2025-02-21 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @olivierverdier
Not really. You have to change coordinates back and forth.
(it's not much, right? either $`x \mapsto x - b`$ (`ShiftedPositiveNumber`), or $`x \mapsto -x -b`$ (`ShiftedNegativeNumber`), or $`x \mapsto -x`$ (`SymmetricNegativeDefinite`). These operations are pretty much for free.).

As long as *these three classes are removed* I have no further objections.

---

You can work with natural coordinates all you want, but the actual point in the manifold may be encoded with a (very simple) change of variable.

The implementation of the exponential distribution could be for instance:

```julia
"""
    get_natural_manifold_base(::Type{Exponential}, ::Tuple{}, conditioner=nothing)

Get the natural manifold base for the `Exponential` distribution.
"""
function get_natural_manifold_base(::Type{Exponential}, ::Tuple{}, conditioner=nothing)
    return PositiveNumbers()
end

"""
    partition_point(::Type{Exponential}, ::Tuple{}, p, conditioner=nothing)

Converts the `point` to a compatible representation for the natural manifold of type `Exponential`.
"""
function partition_point(::Type{Exponential}, ::Tuple{}, p, conditioner=nothing)
    return -p
end
```

If you really need some other manifold for some reason, sure, go ahead, but the three ones that you created should be removed.
