### 2025-02-27 7dce728
fix: erroring generic method
