### 2025-02-27 156ae23
fix: implementation example weibull
