+++ src/natural_manifolds/weibull.jl
@@ -5,7 +5,7 @@ Get the natural manifold base for the `Weibull` distribution.
 """
 function get_natural_manifold_base(::Type{Weibull}, ::Tuple{}, conditioner=nothing)
     @assert conditioner > 0 "Conditioner $(conditioner) should be positive"
-    return ShiftedNegativeNumbers(static(0))
+    return PositiveVectors(1)
 end
 
 """
@@ -15,5 +15,15 @@ Converts the `point` to a compatible representation for the natural manifold of 
 """
 function partition_point(::Type{Weibull}, ::Tuple{}, p, conditioner=nothing)
     @assert conditioner > 0  "Conditioner $(conditioner) should be positive"
+    return -p
+end
+
+"""
+    transform_back!(p, ::NaturalParametersManifold{Weibull}, q)
+
+Transforms the `q` to a compatible representation for the exponential family distribution of type `Weibull`.
+"""
+function transform_back!(p, ::NaturalParametersManifold{ℝ, Weibull}, q)
+    p .= -q
     return p
 end
