### 2025-02-27 009973d
fix: implementation example pareto
