+++ src/natural_manifolds/pareto.jl
@@ -5,7 +5,7 @@
 Get the natural manifold base for the `Pareto` distribution.
 """
 function get_natural_manifold_base(::Type{Pareto}, ::Tuple{}, conditioner=nothing)
-    return ShiftedNegativeNumbers(static(-1))
+    return PositiveVectors(1)
 end
 
 """
@@ -14,5 +14,15 @@ end
 Converts the `point` to a compatible representation for the natural manifold of type `Pareto`.
 """
 function partition_point(::Type{Pareto}, ::Tuple{}, p, conditioner=nothing)
+    return -p .- 1
+end
+
+"""
+    transform_back!(p, ::NaturalParametersManifold{Pareto}, q)
+
+Transforms the `q` to a compatible representation for the exponential family distribution of type `Pareto`.
+"""
+function transform_back!(p, ::NaturalParametersManifold{ℝ, Pareto}, q)
+    p .= -(q .+ 1)
     return p
 end
