+++ src/natural_manifolds/negative_binomial.jl
@@ -17,3 +17,13 @@ function partition_point(::Type{NegativeBinomial}, ::Tuple{}, p, conditioner=not
     @assert conditioner >= 0 "Conditioner should be non-negative"
     return p
 end
+
+"""
+    transform_back!(p, ::NaturalParametersManifold{NegativeBinomial}, q)
+
+Transforms the `q` to a compatible representation for the exponential family distribution of type `NegativeBinomial`.
+"""
+function transform_back!(p, ::NaturalParametersManifold{ℝ, NegativeBinomial}, q)
+    p .= q
+    return p
+end
