+++ test/natural_manifolds/natural_manifolds_setuptests.jl
@@ -17,5 +17,8 @@ function test_natural_manifold(f; seed=42, ndistributions=100)
         η = partition_point(T, dims, getnaturalparameters(ef), getconditioner(ef))
 
         @test is_point(M, η, error=:error)
+        ef_back = convert(ExponentialFamilyDistribution, M, η)
+        @test getnaturalparameters(ef_back) ≈ getnaturalparameters(ef)
+        @test getconditioner(ef_back) == getconditioner(ef)
     end
 end
