+++ src/natural_manifolds/bernoulli.jl
@@ -15,3 +15,13 @@ Converts the `point` to a compatible representation for the natural manifold of 
 function partition_point(::Type{Bernoulli}, ::Tuple{}, p, conditioner=nothing)
     return p
 end
+
+"""
+    transform_back!(p, M::NaturalParametersManifold{ℝ, Bernoulli}, q)
+
+Transforms the natural parameter `q` back to the probability parameter for the Bernoulli distribution.
+"""
+function transform_back!(p, ::NaturalParametersManifold{ℝ, Bernoulli}, q)
+    p .= q
+    return p
+end
