+++ src/natural_manifolds/wishart.jl
@@ -5,7 +5,7 @@ Get the natural manifold base for the `WishartFast` distribution.
 """
 function get_natural_manifold_base(::Type{ExponentialFamily.WishartFast}, dims::Tuple{Int, Int}, conditioner=nothing)
     k = first(dims)
-    return ProductManifold(ShiftedPositiveNumbers(static(0)), SymmetricNegativeDefinite(k))
+    return ProductManifold(PositiveVectors(1), SymmetricPositiveDefinite(k))
 end
 
 """
@@ -15,5 +15,18 @@ Converts the `point` to a compatible representation for the natural manifold of 
 """
 function partition_point(::Type{ExponentialFamily.WishartFast}, dims::Tuple{Int, Int}, p, conditioner=nothing)
     k = first(dims)
-    return ArrayPartition(view(p, 1:1), reshape(view(p, 2:(1 + k^2)), (k, k)))
+    return ArrayPartition(view(p, 1:1), -reshape(view(p, 2:(1 + k^2)), (k, k)))
+end
+
+"""
+    transform_back!(p, M::NaturalParametersManifold{ℝ, WishartFast}, q)
+
+Transforms the `q` to a compatible representation for the exponential family distribution of type `WishartFast`.
+"""
+function transform_back!(p, M::NaturalParametersManifold{ℝ, ExponentialFamily.WishartFast}, q)
+    # k = first(getdims(M))
+    # p[1:1] .= view(q, 1:1)
+    p .= -q
+    p[1:1] .= -view(q, 1:1)
+    return p
 end
