+++ src/natural_manifolds/binomial.jl
@@ -17,3 +17,13 @@ function partition_point(::Type{Binomial}, ::Tuple{}, p, conditioner=nothing)
     @assert conditioner >= 0 "Conditioner $(conditioner) should be positive"
     return p
 end
+
+"""
+    transform_back!(p, ::NaturalParametersManifold{Binomial}, q)
+
+Transforms the `q` to a compatible representation for the exponential family distribution of type `Binomial`.
+"""
+function transform_back!(p, ::NaturalParametersManifold{ℝ, Binomial}, q)
+    p .= q
+    return p
+end
