### 2025-03-13 8375579
fix: try to fix EFP with jacobian_natural_to_canonical
