+++ src/natural_manifolds/negative_binomial.jl
@@ -27,3 +27,10 @@ function transform_back!(p, ::NaturalParametersManifold{ℝ, NegativeBinomial}, 
     p .= q
     return p
 end
+
+function jacobian_natural_to_canonical!(Y, ::NaturalParametersManifold{ℝ, NegativeBinomial}, X)
+    Y .= 1
+    return Y
+end
+
+
