+++ src/natural_manifolds/gamma.jl
@@ -30,3 +30,20 @@ function transform_back!(p, ::NaturalParametersManifold{ℝ, Gamma}, q)
     p[2:2] .= -view(q, 2:2)
     return p
 end
+
+function transform_back(::NaturalParametersManifold{ℝ, Gamma}, q)
+    p = similar(q)
+    p[1:1] .= view(q, 1:1) .- 1
+    p[2:2] .= -view(q, 2:2)
+    return p
+end
+
+function jacobian_natural_to_canonical(::NaturalParametersManifold{ℝ, Gamma}, X)
+    # For Gamma, the Jacobian is diagonal with [1, -1]
+    # This means the first component of the gradient stays the same
+    # But the second component changes sign
+    result = similar(X)
+    result[1:1] .= X[1:1]  # First component stays the same
+    result[2:2] .= -X[2:2]  # Second component changes sign
+    return result
+end
