+++ src/natural_manifolds/inverse_gamma.jl
@@ -3,7 +3,9 @@
 
 Get the natural manifold base for the `ExponentialFamily.GammaInverse` distribution.
 """
-function get_natural_manifold_base(::Type{ExponentialFamily.GammaInverse}, ::Tuple{}, conditioner=nothing)
+function get_natural_manifold_base(
+    ::Type{ExponentialFamily.GammaInverse}, ::Tuple{}, conditioner=nothing
+)
     return ProductManifold(
         ShiftedNegativeNumbers(static(-1)), ShiftedNegativeNumbers(static(0))
     )
@@ -14,6 +16,8 @@ end
 
 Converts the `point` to a compatible representation for the natural manifold of type `ExponentialFamily.GammaInverse`.
 """
-function partition_point(::Type{ExponentialFamily.GammaInverse}, ::Tuple{}, p, conditioner=nothing)
+function partition_point(
+    ::Type{ExponentialFamily.GammaInverse}, ::Tuple{}, p, conditioner=nothing
+)
     return ArrayPartition(view(p, 1:1), view(p, 2:2))
 end
