### 2025-03-17 e1daf80
feat: implement retraction selection mechanism
