+++ src/natural_manifolds.jl
@@ -8,20 +8,30 @@ import ExponentialFamily: exponential_family_typetag
 The manifold for the natural parameters of the distribution of type `T` with dimensions `dims`.
 An internal structure, use `get_natural_manifold` to create an instance of a manifold for the natural parameters of distribution of type `T`.
 """
-struct NaturalParametersManifold{𝔽,T,D,M,C} <: AbstractDecoratorManifold{𝔽}
+struct NaturalParametersManifold{𝔽,T,D,M,C,R} <: AbstractDecoratorManifold{𝔽}
     dims::D
     base::M
     conditioner::C
+    retraction::R
 end
 
 getdims(M::NaturalParametersManifold) = M.dims
 getbase(M::NaturalParametersManifold) = M.base
 getconditioner(M::NaturalParametersManifold) = M.conditioner
+getretraction(M::NaturalParametersManifold) = M.retraction
 
 # The `NaturalParametersManifold` simply adds extra properties to the `base` and 
 # acts as a "decorator"
-@inline ManifoldsBase.active_traits(f::F, ::NaturalParametersManifold, ::Any...) where {F} =
-    ManifoldsBase.IsExplicitDecorator()
+@inline function ManifoldsBase.active_traits(f::F, ::NaturalParametersManifold, args...) where {F}
+    # Don't delegate retraction-related methods
+    if f in (ManifoldsBase.retract, ManifoldsBase.retract!, 
+             ManifoldsBase.retract_fused, ManifoldsBase.retract_fused!)
+        return ManifoldsBase.EmptyTrait()
+    else
+        return ManifoldsBase.IsExplicitDecorator()
+    end
+end
+
 @inline ManifoldsBase.decorated_manifold(M::NaturalParametersManifold) = M.base
 
 function ExponentialFamily.exponential_family_typetag(
@@ -31,9 +41,9 @@ function ExponentialFamily.exponential_family_typetag(
 end
 
 function NaturalParametersManifold(
-    ::Type{T}, dims::D, base::M, conditioner::C=nothing
-) where {T,𝔽,D,M<:AbstractManifold{𝔽},C}
-    return NaturalParametersManifold{𝔽,T,D,M,C}(dims, base, conditioner)
+    ::Type{T}, dims::D, base::M, conditioner::C=nothing, retraction::R=nothing
+) where {T,𝔽,D,M<:AbstractManifold{𝔽},C,R}
+    return NaturalParametersManifold{𝔽,T,D,M,C,R}(dims, base, conditioner, retraction)
 end
 
 """
@@ -52,9 +62,9 @@ julia> ExponentialFamilyManifolds.get_natural_manifold(MvNormalMeanCovariance, (
 true
 ```
 """
-function get_natural_manifold(::Type{T}, dims, conditioner=nothing) where {T}
+function get_natural_manifold(::Type{T}, dims, conditioner=nothing, retraction = nothing) where {T}
     return NaturalParametersManifold(
-        T, dims, get_natural_manifold_base(T, dims, conditioner), conditioner
+        T, dims, get_natural_manifold_base(T, dims, conditioner), conditioner, retraction
     )
 end
 
@@ -88,3 +98,28 @@ function Base.convert(
         exponential_family_typetag(M), p, getconditioner(M), nothing
     )
 end
+
+struct ChartNOrderRetraction{Order, E} <: AbstractRetractionMethod
+    extra::E
+end
+
+function ChartNOrderRetraction{O}() where {O}
+    return ChartNOrderRetraction{O, Nothing}(nothing)
+end
+
[…]
