### 2025-03-17 582e0bd
feat: draft Fisher Rao metric
