+++ src/natural_manifolds.jl
@@ -1,6 +1,30 @@
 using ManifoldsBase, Manifolds, Static, RecursiveArrayTools, Random, ExponentialFamily
+using FastCholesky
 
 import ExponentialFamily: exponential_family_typetag
+
+"""
+    FisherInformationMetric <: RiemannianMetric
+
+Specifier that we need to use the Fisher information metric.
+"""
+struct FisherInformationMetric <: RiemannianMetric end
+
+"""
+    BaseMetric <: RiemannianMetric
+
+Specifier that we need to use the metric from the base manifold.
+"""
+struct BaseMetric <: RiemannianMetric end
+
+"""
+    getdefaultmetric(::Type{T}) where {T}
+
+Returns the default metric for the distribution of type `T`.
+"""
+function getdefaultmetric(::Type{T}) where {T}
+    return FisherInformationMetric()
+end
 
 """
     NaturalParametersManifold(::Type{T}, dims, base, conditioner)
@@ -8,11 +32,12 @@ import ExponentialFamily: exponential_family_typetag
 The manifold for the natural parameters of the distribution of type `T` with dimensions `dims`.
 An internal structure, use `get_natural_manifold` to create an instance of a manifold for the natural parameters of distribution of type `T`.
 """
-struct NaturalParametersManifold{𝔽,T,D,M,C,R} <: AbstractDecoratorManifold{𝔽}
+struct NaturalParametersManifold{𝔽,T,D,M,C,R,MT} <: AbstractDecoratorManifold{𝔽}
     dims::D
     base::M
     conditioner::C
     retraction::R
+    metric::MT
 end
 
 getdims(M::NaturalParametersManifold) = M.dims
@@ -22,20 +47,29 @@ getretraction(M::NaturalParametersManifold) = M.retraction
 
 # The `NaturalParametersManifold` simply adds extra properties to the `base` and 
 # acts as a "decorator"
-@inline function ManifoldsBase.active_traits(
-    f::F, ::NaturalParametersManifold, args...
-) where {F}
-    # Don't delegate retraction-related methods
+function select_skip_methods(::F, ::NaturalParametersManifold{𝔽,T,D,MB,C,R,BaseMetric}) where {F,𝔽,T,D,MB,C,R}
+    return ManifoldsBase.IsExplicitDecorator()
+end
+
+function select_skip_methods(f::F, ::NaturalParametersManifold{𝔽,T,D,MB,C,R,FisherInformationMetric}) where {F,𝔽,T,D,MB,C,R}
     if f in (
         ManifoldsBase.retract,
         ManifoldsBase.retract!,
         ManifoldsBase.retract_fused,
         ManifoldsBase.retract_fused!,
+        Manifolds.local_metric,
+        Manifolds.local_metric_jacobian,
     )
         return ManifoldsBase.EmptyTrait()
     else
         return ManifoldsBase.IsExplicitDecorator()
     end
+end
+
+@inline function ManifoldsBase.active_traits(
+    f::F, M::NaturalParametersManifold, args...
+) where {F}
+    return select_skip_methods(f, M)
 end
 
 @inline ManifoldsBase.decorated_manifold(M::NaturalParametersManifold) = M.base
@@ -47,9 +81,16 @@ function ExponentialFamily.exponential_family_typetag(
 end
 
 function NaturalParametersManifold(
[…]
