+++ test/natural_manifolds/natural_manifolds_setuptests.jl
@@ -1,9 +1,9 @@
 
-using StableRNGs, ExponentialFamily, ManifoldsBase, LinearAlgebra
+using StableRNGs, ExponentialFamily, Manifolds, ManifoldsBase, LinearAlgebra
 
 import ExponentialFamilyManifolds: get_natural_manifold, partition_point
 
-function test_natural_manifold(f; seed=42, ndistributions=100)
+function test_natural_manifold(f; seed=42, ndistributions=100, test_metric=true)
     rng = StableRNG(seed)
 
     foreach(1:ndistributions) do _
@@ -17,5 +17,11 @@ function test_natural_manifold(f; seed=42, ndistributions=100)
         η = partition_point(T, dims, getnaturalparameters(ef), getconditioner(ef))
 
         @test is_point(M, η, error=:error)
+
+        if test_metric
+            @test M.metric isa ExponentialFamilyManifolds.FisherInformationMetric
+            @test ManifoldsBase.get_basis_default(M, η) isa ExponentialFamilyManifolds.NaturalBasis
+            @test Manifolds.local_metric(M, η, ManifoldsBase.get_basis_default(M, η)) ≈ ExponentialFamily.fisherinformation(ef)
+        end
     end
 end
