+++ test/natural_manifolds/normal_tests.jl
@@ -9,13 +9,21 @@ end
 @testitem "Check `MvNormal` natural manifold" begin
     include("natural_manifolds_setuptests.jl")
 
-    test_natural_manifold() do rng
+    test_natural_manifold(; test_metric=false) do rng
         k = rand(rng, 1:10)
         m = randn(k)
         L = LowerTriangular(randn(k, k))
         C = L * L' + k * I
         return MvNormalMeanCovariance(m, C)
     end
+
+    using ExponentialFamilyManifolds
+    using ExponentialFamily
+
+    @test ExponentialFamilyManifolds.getdefaultmetric(MvNormalMeanCovariance) isa
+        ExponentialFamilyManifolds.BaseMetric
+    M = ExponentialFamilyManifolds.get_natural_manifold(MvNormalMeanCovariance, (3,))
+    @test M.metric isa ExponentialFamilyManifolds.BaseMetric
 end
 
 @testitem "Check `MvNormalMeanScalePrecision` natural manifold" begin
