+++ Project.toml
@@ -6,6 +6,7 @@ version = "2.0.0"
 [deps]
 BayesBase = "b4ee3484-f114-42fe-b91c-797d54a0c67e"
 ExponentialFamily = "62312e5e-252a-4322-ace9-a5f4bf9b357b"
+FastCholesky = "2d5283b6-8564-42b6-bb00-83ed8e915756"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Manifolds = "1cead3c2-87b3-11e9-0ccd-23c62b72b94e"
 ManifoldsBase = "3362f125-f0bb-47a3-aa74-596ffd7ef2fb"
@@ -13,10 +14,15 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 RecursiveArrayTools = "731186ca-8d62-57ce-b412-fbd966d074cd"
 Static = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
 
+[weakdeps]
+ManifoldDiff = "af67fdf4-a580-4b9f-bbec-742ef357defd"
+
 [compat]
 BayesBase = "1.3"
 ExponentialFamily = "2.0.0"
+FastCholesky = "1.3.1"
 LinearAlgebra = "1.10"
+ManifoldDiff = "0.4.2"
 Manifolds = "0.10"
 ManifoldsBase = "1"
 Random = "1.10"
