### 2025-03-17 a1d7584
feat: draft generic SecondOrderRetraction
