+++ Project.toml
@@ -18,6 +18,7 @@ Static = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
 ManifoldDiff = "af67fdf4-a580-4b9f-bbec-742ef357defd"
 
 [compat]
+ADTypes = "1.14.0"
 BayesBase = "1.3"
 ExponentialFamily = "2.0.0"
 FastCholesky = "1.3.1"
@@ -31,6 +32,7 @@ Static = "0.8, 1"
 julia = "1.10"
 
 [extras]
+ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
@@ -42,4 +44,4 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Aqua", "Distributions", "JET", "Test", "ReTestItems", "StableRNGs", "StaticArrays", "Manopt", "ForwardDiff"]
+test = ["ADTypes", "Aqua", "Distributions", "JET", "Test", "ReTestItems", "StableRNGs", "StaticArrays", "Manopt", "ForwardDiff"]
