### 2025-03-18 4d6fad4
refactor: use metric to select retraction method
