+++ src/natural_manifolds/normal.jl
@@ -67,13 +67,6 @@ function partition_point(
     return ArrayPartition(view(p, 1:k), view(p, (k + 1):(k + 1)))
 end
 
-function ManifoldsBase.default_retraction_method(
-    ::NaturalParametersManifold{𝔽,MvNormalMeanCovariance,D,M,C,Nothing,BaseMetric},
-    ::Type{T},
-) where {𝔽,T,D,M,C}
-    return ManifoldsBase.ExponentialRetraction()
-end
-
 function getdefaultmetric(::Type{MvNormalMeanCovariance})
     return BaseMetric()
 end
