+++ src/natural_manifolds/gamma.jl
@@ -18,3 +18,5 @@ Converts the `point` to a compatible representation for the natural manifold of 
 function partition_point(::Type{Gamma}, ::Tuple{}, p, conditioner=nothing)
     return ArrayPartition(view(p, 1:1), view(p, 2:2))
 end
+
+Manifolds.representation_size(::NaturalParametersManifold{𝔽, Gamma}) where {𝔽} = (2,)
