+++ test/natural_manifolds/categorical_tests.jl
@@ -1,7 +1,7 @@
 @testitem "Check `Categorical` natural manifold" begin
     include("natural_manifolds_setuptests.jl")
 
-    test_natural_manifold() do rng
+    test_natural_manifold(test_injectivity_radius=false) do rng
         p = rand(rng, 10)
         normalize!(p, 1)
         return Categorical(p)
