### 2025-03-18 621880e
test: fisher of MvNormalMeanScalePrecision is not fd friendly
