+++ test/natural_manifolds/bernoulli_tests.jl
@@ -5,3 +5,14 @@
         return Bernoulli(rand(rng))
     end
 end
+
+@testitem "Check MLE works for `Bernoulli`" begin
+    include("natural_manifolds_setuptests.jl")
+
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    test_mle_works(mle_samples=1000, ndistributions=3) do rng
+        return Bernoulli(rand(rng))
+    end
+end
