+++ test/natural_manifolds/rayleigh_tests.jl
@@ -5,3 +5,13 @@
         return Rayleigh(10rand(rng))
     end
 end
+
+@testitem "Check MLE works for `Rayleigh`" begin
+    include("natural_manifolds_setuptests.jl")
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    test_mle_works(mle_samples=500, ndistributions=3) do rng
+        return Rayleigh(rand(rng))
+    end
+end
