+++ test/natural_manifolds/gamma_inverse_tests.jl
@@ -5,3 +5,14 @@
         return ExponentialFamily.InverseGamma(10rand(rng), 10rand(rng))
     end
 end
+
+@testitem "Check MLE works for `GammaInverse`" begin
+    include("natural_manifolds_setuptests.jl")
+
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    test_mle_works(mle_samples=500, ndistributions=3) do rng
+        return ExponentialFamily.InverseGamma(10rand(rng), 10rand(rng))
+    end
+end
