+++ test/natural_manifolds/chisq_tests.jl
@@ -5,3 +5,14 @@
         return Chisq(10rand(rng))
     end
 end
+
+@testitem "Check MLE works for `Chisq`" begin
+    include("natural_manifolds_setuptests.jl")
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    # Use fewer samples/iterations for faster tests
+    test_mle_works(mle_samples=200, ndistributions=3) do rng
+        return Chisq(10rand(rng))
+    end
+end
