+++ test/natural_manifolds/pareto_tests.jl
@@ -5,3 +5,13 @@
         return Pareto(10rand(rng), 10rand(rng))
     end
 end
+
+@testitem "Check MLE works for `Pareto`" begin
+    include("natural_manifolds_setuptests.jl")
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    test_mle_works(mle_samples=500, ndistributions=3) do rng
+        return Pareto(rand(rng), rand(rng))
+    end
+end
