+++ test/natural_manifolds/poisson_tests.jl
@@ -5,3 +5,14 @@
         return Poisson(rand(rng))
     end
 end
+
+@testitem "Check MLE works for `Poisson`" begin
+    include("natural_manifolds_setuptests.jl")
+    using Manopt
+    import Distributions: kldivergence, Distribution
+
+    # Use fewer samples/iterations for faster tests
+    test_mle_works(mle_samples=500, ndistributions=3) do rng
+        return Poisson(rand(rng))
+    end
+end
