### 2025-03-19 cb1b998
fix: update BayesBase for InvArrowHead
