+++ Project.toml
@@ -19,7 +19,7 @@ ManifoldDiff = "af67fdf4-a580-4b9f-bbec-742ef357defd"
 
 [compat]
 ADTypes = "1.14.0"
-BayesBase = "1.3"
+BayesBase = "1.5.4"
 ExponentialFamily = "2.0.0"
 FastCholesky = "1.3.1"
 LinearAlgebra = "1.10"
