+++ test/natural_manifolds/chisq_tests.jl
@@ -12,7 +12,7 @@ end
     import Distributions: kldivergence, Distribution
 
     # Use fewer samples/iterations for faster tests
-    test_mle_works(mle_samples=200, ndistributions=3) do rng
+    test_mle_works(; mle_samples=200, ndistributions=3) do rng
         return Chisq(10rand(rng))
     end
 end
