+++ test/natural_manifolds/negative_binomial_tests.jl
@@ -11,7 +11,7 @@ end
     using Manopt
     import Distributions: kldivergence, Distribution
 
-    @test_broken test_mle_works(mle_samples=1000, ndistributions=3) do rng
+    @test_broken test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return NegativeBinomial(rand(rng), rand(rng, 0.00001:0.01:1))
     end
 end
