+++ test/natural_manifolds/gamma_tests.jl
@@ -11,7 +11,7 @@ end
     using Manopt
     import Distributions: kldivergence, Distribution
 
-    @test_broken test_mle_works(mle_samples=500, ndistributions=3) do rng
+    @test_broken test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Gamma(rand(rng), 10rand(rng))
     end
 end
