+++ test/natural_manifolds/gamma_inverse_tests.jl
@@ -12,7 +12,7 @@ end
     using Manopt
     import Distributions: kldivergence, Distribution
 
-    test_mle_works(mle_samples=500, ndistributions=3) do rng
+    test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return ExponentialFamily.InverseGamma(10rand(rng), 10rand(rng))
     end
 end
