+++ test/natural_manifolds/dirichlet_tests.jl
@@ -13,7 +13,7 @@ end
     import Distributions: kldivergence, Distribution
 
     # Use fewer samples/iterations for faster tests
-    test_mle_works(mle_samples=500, ndistributions=3) do rng
+    test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Dirichlet(10rand(rng, 3))
     end
 end
