+++ test/natural_manifolds/beta_tests.jl
@@ -12,7 +12,7 @@ end
     import Distributions: kldivergence, Distribution
 
     # Use fewer samples/iterations for faster tests
-    test_mle_works(mle_samples=1000, ndistributions=3) do rng
+    test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return Beta(10rand(rng), 10rand(rng))
     end
 end
