### 2025-03-24 b277f86
fix: use gradient from ManifoldDiff
