+++ test/natural_manifolds/normal_tests.jl
@@ -30,19 +30,14 @@ end
 end
 
 @testitem "Check MLE works for `MvNormalMeanScalePrecision`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     test_mle_works(; mle_samples=1000, ndistributions=3) do rng
         return MvNormalMeanScalePrecision(randn(rng, 2), 2)
     end
 end
 
 @testitem "Check MLE works for `NormalMeanVariance`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
+    include("mle_manifolds_setuptests.jl")
 
     test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return NormalMeanVariance(randn(rng), 1 / 2)
@@ -50,10 +45,7 @@ end
 end
 
 @testitem "Check MLE works for `MvNormalMeanCovariance`" begin
-    include("natural_manifolds_setuptests.jl")
-    using Manopt
-    import Distributions: kldivergence, Distribution
-
+    include("mle_manifolds_setuptests.jl")
     @test_broken test_mle_works(;
         mle_samples=500, ndistributions=3, kl_friendly=false
     ) do rng
