### 2025-03-24 757ad53
fix: almost on all manifolds MLE works
