+++ src/natural_manifolds/chisq.jl
@@ -5,7 +5,7 @@
 Get the natural manifold base for the `Chisq` distribution.
 """
 function get_natural_manifold_base(::Type{Chisq}, ::Tuple{}, conditioner=nothing)
-    return PositiveVectors(1)
+    return ProductManifold(PositiveVectors(1))
 end
 
 """
@@ -14,7 +14,7 @@ end
 Converts the `point` to a compatible representation for the natural manifold of type `Chisq`.
 """
 function partition_point(::Type{Chisq}, ::Tuple{}, p, conditioner=nothing)
-    return p .+ 1
+    return ArrayPartition(view(p, 1:1) .+ 1)
 end
 
 """
@@ -22,7 +22,7 @@ end
 
 Transforms the `q` to a compatible representation for the exponential family distribution of type `Chisq`.
 """
-function transform_back!(p, ::NaturalParametersManifold{ℝ, Chisq}, q)
+function transform_back!(p, ::NaturalParametersManifold{ℝ,Chisq}, q)
     p .= q .- 1
     return p
 end
