+++ src/natural_manifolds/exponential.jl
@@ -4,7 +4,7 @@
 Get the natural manifold base for the `Exponential` distribution.
 """
 function get_natural_manifold_base(::Type{Exponential}, ::Tuple{}, conditioner=nothing)
-    return PositiveVectors(1)
+    return ProductManifold(PositiveVectors(1))
 end
 
 """
@@ -13,7 +13,7 @@ end
 Converts the `point` to a compatible representation for the natural manifold of type `Exponential`.
 """
 function partition_point(::Type{Exponential}, ::Tuple{}, p, conditioner=nothing)
-    return -p
+    return ArrayPartition(-p)
 end
 
 """
@@ -21,7 +21,7 @@ end
 
 Transforms the `q` to a compatible representation for the exponential family distribution of type `Exponential`.
 """
-function transform_back!(p, ::NaturalParametersManifold{ℝ, Exponential}, q)
+function transform_back!(p, ::NaturalParametersManifold{ℝ,Exponential}, q)
     p .= -q
     return p
 end
