+++ src/natural_manifolds/wishart.jl
@@ -27,7 +27,7 @@ end
 
 Transforms the `q` to a compatible representation for the exponential family distribution of type `WishartFast`.
 """
-function transform_back!(p, ::NaturalParametersManifold{ℝ, ExponentialFamily.WishartFast}, q)
+function transform_back!(p, ::NaturalParametersManifold{ℝ,ExponentialFamily.WishartFast}, q)
     p .= -q
     p[1:1] .= view(q, 1:1)
     return p
