+++ test/natural_manifolds/gamma_tests.jl
@@ -9,7 +9,7 @@ end
 @testitem "Check MLE works for `Gamma`" begin
     include("mle_manifolds_setuptests.jl")
 
-    @test_broken test_mle_works(; mle_samples=500, ndistributions=3) do rng
+    test_mle_works(; mle_samples=500, ndistributions=3) do rng
         return Gamma(rand(rng), rand(rng))
     end
 end
