+++ test/natural_manifolds/wishart_tests.jl
@@ -13,11 +13,10 @@ end
     include("mle_manifolds_setuptests.jl")
 
     import ExponentialFamily: WishartFast
+    using DifferentiationInterface
+    using FiniteDifferences
 
-    test_mle_works(; mle_samples=500, ndistributions=1, kl_friendly=false) do rng
-        k = 2
-        L = LowerTriangular(randn(rng, k, k))
-        C = L * L' + k * I
-        return WishartFast(k + 2, C)
+    test_mle_works(; mle_samples=500, ndistributions=1, backend_type=AutoFiniteDifferences(central_fdm(5, 1)), kl_friendly=false) do _
+        return WishartFast(4, diagm([1, 1]))
     end
 end
