+++ Project.toml
@@ -16,7 +16,9 @@ Static = "aedffcd0-7271-4cad-89d0-dc628f76c6d3"
 [compat]
 ADTypes = "1.14.0"
 BayesBase = "1.3"
+DifferentiationInterface = "0.6.48"
 ExponentialFamily = "2.0.0"
+FiniteDifferences = "0.12.32"
 LinearAlgebra = "1.10"
 ManifoldDiff = "0.4.2"
 Manifolds = "0.10"
@@ -29,15 +31,17 @@ julia = "1.10"
 [extras]
 ADTypes = "47edcb42-4c32-4615-8424-f2b9edc5f35b"
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
+FiniteDifferences = "26cc04aa-876d-5657-8c51-4c34ba976000"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
-Manopt = "0fc0a36d-df90-57f3-8f93-d78a9fc72bb5"
 ManifoldDiff = "af67fdf4-a580-4b9f-bbec-742ef357defd"
+Manopt = "0fc0a36d-df90-57f3-8f93-d78a9fc72bb5"
 ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["ADTypes", "Aqua", "Distributions", "JET", "Test", "ReTestItems", "StableRNGs", "StaticArrays", "ManifoldDiff", "Manopt", "ForwardDiff"]
+test = ["ADTypes", "Aqua", "DifferentiationInterface", "Distributions", "JET", "Test", "ReTestItems", "StableRNGs", "StaticArrays", "ManifoldDiff", "Manopt", "ForwardDiff", "FiniteDifferences"]
