+++ src/symmetric_negative_definite.jl
@@ -1,207 +1,0 @@
-"""
-    SymmetricNegativeDefinite(k)
-
-This manifold represents the set of negative definite matrices of size `k × k`. 
-Similar to `SymmetricPositiveDefinite` from `Manifolds.jl` with the exception that the matrices are negative definite.
-"""
-struct SymmetricNegativeDefinite{B} <: AbstractManifold{ℝ}
-    base::B
-
-    function SymmetricNegativeDefinite(size)
-        base = SymmetricPositiveDefinite(size)
-        return new{typeof(base)}(base)
-    end
-end
-
-function Base.show(io::IO, M::SymmetricNegativeDefinite)
-    return print(io, "SymmetricNegativeDefinite(", first(representation_size(M.base)), ")")
-end
-
-function ManifoldsBase.get_embedding(M::SymmetricNegativeDefinite)
-    return ManifoldsBase.get_embedding(M.base)
-end
-
-function ManifoldsBase.representation_size(M::SymmetricNegativeDefinite)
-    return ManifoldsBase.representation_size(M.base)
-end
-function ManifoldsBase.manifold_dimension(M::SymmetricNegativeDefinite)
-    return ManifoldsBase.manifold_dimension(M.base)
-end
-function ManifoldsBase.injectivity_radius(M::SymmetricNegativeDefinite)
-    return ManifoldsBase.injectivity_radius(M.base)
-end
-ManifoldsBase.is_flat(M::SymmetricNegativeDefinite) = ManifoldsBase.is_flat(M.base)
-
-function ManifoldsBase.check_point(M::SymmetricNegativeDefinite, p; kwargs...)
-    if !isapprox(norm(p - transpose(p)), 0.0; kwargs...)
-        return DomainError(
-            norm(p - transpose(p)),
-            lazy"The point $(p) does not lie on $(M) since its not a symmetric matrix:",
-        )
-    end
-    if !isposdef(Negated(p))
-        return DomainError(
-            eigvals(p),
-            lazy"The point $p does not lie on $(M) since its not a negative definite matrix.",
-        )
-    end
-    return nothing
-end
-
-function ManifoldsBase.check_vector(M::SymmetricNegativeDefinite, p, X; kwargs...)
-    if !isapprox(X, transpose(X); kwargs...)
-        return DomainError(
-            X,
-            lazy"The vector $(X) is not a tangent to a point on $(M) (represented as an element of the Lie algebra) since its not symmetric.",
-        )
-    end
-    return nothing
-end
-
-ManifoldsBase.embed(::SymmetricNegativeDefinite, p) = p
-ManifoldsBase.embed(::SymmetricNegativeDefinite, p, X) = X
-
-function ManifoldsBase.inner(M::SymmetricNegativeDefinite, p, X, Y)
-    return ManifoldsBase.inner(M.base, Negated(p), Negated(X), Negated(Y))
-end
-
-function ManifoldsBase.exp_fused!(M::SymmetricNegativeDefinite, q, p, X, t::Number)
-    ManifoldsBase.exp_fused!(M.base, q, Negated(p), Negated(X), t)
-    return negate!(q)
-end
-
-function ManifoldsBase.exp!(M::SymmetricNegativeDefinite, q, p, X)
-    return ManifoldsBase.exp_fused!(M, q, p, X, one(eltype(p)))
-end
-
-function ManifoldsBase.log!(M::SymmetricNegativeDefinite, X, p, q)
-    ManifoldsBase.log!(M.base, X, Negated(p), Negated(q))
[…]
