### 2025-03-24 5f5b2a0
docs: use appropriate gradient in example
