### 2025-03-24 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @Nimrais
@olivierverdier Hi, sorry for the month delay, but this package is a foundation package for many packages in the ecosystem, so changes here broke many other packages. I needed to prepare changes, or at least the possibility of their implementation, beforehand. Your main concern is addressed in this PR. If you agree, you can approve the PR and close the issue once the PR is merged.

But anyway, thanks for such a careful review! Your suggestion turned out to be a really good one!
