### 2025-03-27 4613d14
docs: update paper example accordingly to changes
