+++ paper_example/paper_example.jl
@@ -2,33 +2,50 @@ using Distributions, LinearAlgebra, Random, Plots
 using ExponentialFamily, ExponentialFamilyManifolds
 using Manopt, ForwardDiff, StableRNGs
 using PGFPlotsX, RecursiveArrayTools
+using ManifoldDiff, DifferentiationInterface
+using ManifoldDiff: TangentDiffBackend
+using FiniteDifferences
+
+"""
+Run optimization multiple times with different random initializations
+and return the best result based on the objective function value.
+"""
+function optimize_multiple_starts(M, f, n_starts=10; rng=StableRNG(42))
+    best_p = nothing
+    best_value = Inf
+    
+    for _ in 1:n_starts
+        p_init = rand(rng, M)
+        p_opt = optimize(M, f, p_init)
+        current_value = f(M, p_opt)
+        
+        if current_value < best_value
+            best_value = current_value
+            best_p = p_opt
+        end
+    end
+    
+    return convert(Distribution, convert(ExponentialFamilyDistribution, M, best_p))
+end
 
 """
 Generic optimization function as shown in Listing 1 of the paper.
 Performs gradient descent optimization on manifold M with objective f.
 """
-function optimize(M, f, η_init; max_iters=500)
-    g(M, η) = ForwardDiff.gradient(
-        x -> f(M, x), η
-    )
-    
-    stopping_criterion = StopAfterIteration(max_iters) | 
-                        StopWhenGradientNormLess(1e-6)
-    
-    η_opt = gradient_descent(M, f, g, η_init;
-                           stopping_criterion=stopping_criterion)
-    
-    return convert(Distribution, 
-           convert(ExponentialFamilyDistribution, M, η_opt))
+function optimize(M, f, p_init)
+    autograd_backend = AutoFiniteDifferences(central_fdm(3, 1))
+    backend = TangentDiffBackend(autograd_backend)
+    grad = (M, p) -> ManifoldDiff.gradient(M, (p) -> f(M, p), p, backend)
+    p_opt = gradient_descent(M, f, grad, p_init)
+    return p_opt
 end
 
 """
 MLE objective function as shown in Listing 2 of the paper.
 """
 function f_mle(M, η, samples)
-    dist = convert(ExponentialFamilyDistribution, M, η)
-    dist = convert(Distribution, dist)
-    return -mean(logpdf(dist, s) for s in eachcol(samples))
+    ef = convert(ExponentialFamilyDistribution, M, η)
+    return -mean(logpdf(ef, s) for s in eachcol(samples))
 end
 
 """
@@ -41,9 +58,8 @@ function f_vi(M, η, target_dist; n_samples=2000)
     # Sample from q using stable RNG for reproducibility
     samples = rand(StableRNG(22), q_dist, n_samples)
     
-    # Compute difference of log probabilities
-    diff(s) = logpdf(target_dist, s) - logpdf(q_dist, s)
-    return -mean(diff(samples[:, i]) for i in 1:n_samples)
+    diff(s) = logpdf(q_dist, s) - logpdf(target_dist, s)
+    return mean(diff(samples[:, i]) for i in 1:n_samples)
 end
 
 """
@@ -63,7 +79,7 @@ Creates a mixture of two Gaussians as target distribution
 for demonstrating the differences between MLE and VI.
 """
 function create_target_distribution()
-    μ1 = [0.9, 1.0]
+    μ1 = [3.5, 3.5]
[…]
