+++ test/natural_manifolds_tests.jl
@@ -87,7 +87,9 @@ end
 
     ef = ExponentialFamilyDistribution(CustomDistribution, ())
 
-    function ExponentialFamilyManifolds.get_natural_manifold_base(::Type{CustomDistribution}, ::Tuple{}, conditioner=nothing)
+    function ExponentialFamilyManifolds.get_natural_manifold_base(
+        ::Type{CustomDistribution}, ::Tuple{}, conditioner=nothing
+    )
         return Euclidean(1)
     end
 
@@ -96,5 +98,12 @@ end
     q = rand(M)
 
     @test_throws ErrorException ExponentialFamilyManifolds.transform_back!(p, M, q)
-    @test occursin("You need to implement `transform_back!` for your specific", try ExponentialFamilyManifolds.transform_back!(p, M, q) catch e e.msg end)
+    @test occursin(
+        "You need to implement `transform_back!` for your specific",
+        try
+            ExponentialFamilyManifolds.transform_back!(p, M, q)
+        catch e
+            e.msg
+        end,
+    )
 end
