+++ paper_example/paper_example.jl
@@ -13,18 +13,18 @@ and return the best result based on the objective function value.
 function optimize_multiple_starts(M, f, n_starts=10; rng=StableRNG(42))
     best_p = nothing
     best_value = Inf
-    
+
     for _ in 1:n_starts
         p_init = rand(rng, M)
         p_opt = optimize(M, f, p_init)
         current_value = f(M, p_opt)
-        
+
         if current_value < best_value
             best_value = current_value
             best_p = p_opt
         end
     end
-    
+
     return convert(Distribution, convert(ExponentialFamilyDistribution, M, best_p))
 end
 
@@ -54,10 +54,10 @@ VI objective function (Free Energy) as shown in Listing 3 of the paper.
 function f_vi(M, η, target_dist; n_samples=2000)
     q = convert(ExponentialFamilyDistribution, M, η)
     q_dist = convert(Distribution, q)
-    
+
     # Sample from q using stable RNG for reproducibility
     samples = rand(StableRNG(22), q_dist, n_samples)
-    
+
     diff(s) = logpdf(q_dist, s) - logpdf(target_dist, s)
     return mean(diff(samples[:, i]) for i in 1:n_samples)
 end
@@ -68,9 +68,7 @@ Natural gradient implementation as shown in Listing 4 of the paper.
 function natural_gradient(M, η, f)
     q = convert(ExponentialFamilyDistribution, M, η)
     F = fisherinformation(q)
-    grad = ForwardDiff.gradient(
-        x -> f(M, x), η
-    )
+    grad = ForwardDiff.gradient(x -> f(M, x), η)
     return F \ grad
 end
 
@@ -81,15 +79,12 @@ for demonstrating the differences between MLE and VI.
 function create_target_distribution()
     μ1 = [3.5, 3.5]
     Σ1 = [1.0 0.2; 0.2 1.2]
-    
+
     μ2 = [-1.3, -1]
     Σ2 = [1.2 0.3; 0.3 1.5]
-    
-    mixture = MixtureModel([
-        MvNormal(μ1, Σ1),
-        MvNormal(μ2, Σ2)
-    ], [0.4, 0.6])
-    
+
+    mixture = MixtureModel([MvNormal(μ1, Σ1), MvNormal(μ2, Σ2)], [0.4, 0.6])
+
     return mixture
 end
 
@@ -98,22 +93,42 @@ Helper function to visualize and compare two different approximations
 to the target distribution.
 """
 function plot_comparison(q1, q2, target_dist, title_q1, title_q2)
-    x = range(-8, 8, length=100)
-    y = range(-8, 8, length=100)
+    x = range(-8, 8; length=100)
+    y = range(-8, 8; length=100)
 
     z_target = [pdf(target_dist, [xi, yi]) for yi in y, xi in x]
     z_p = [pdf(q1, [xi, yi]) for yi in y, xi in x]
     z_q = [pdf(q2, [xi, yi]) for yi in y, xi in x]
 
-    plt1 = contour(x, y, z_target; fill=true, alpha=0.4, color=:viridis,
-        levels=15, title=title_q1, xlabel="x1", ylabel="x2")
+    plt1 = contour(
+        x,
+        y,
+        z_target;
+        fill=true,
+        alpha=0.4,
+        color=:viridis,
+        levels=15,
+        title=title_q1,
+        xlabel="x1",
+        ylabel="x2",
+    )
     contour!(plt1, x, y, z_p; color=:red, fill=false, linewidth=2, levels=10)
 
-    plt2 = contour(x, y, z_target; fill=true, alpha=0.4, color=:viridis,
[…]
