+++ test/natural_manifolds/wishart_tests.jl
@@ -16,7 +16,12 @@ end
     using DifferentiationInterface
     using FiniteDifferences
 
-    test_mle_works(; mle_samples=500, ndistributions=1, backend_type=AutoFiniteDifferences(central_fdm(5, 1)), kl_friendly=false) do _
+    test_mle_works(;
+        mle_samples=500,
+        ndistributions=1,
+        backend_type=AutoFiniteDifferences(central_fdm(5, 1)),
+        kl_friendly=false,
+    ) do _
         return WishartFast(4, diagm([1, 1]))
     end
 end
