### 2025-03-27 18df1d6
fix: update paper example folder
