+++ paper_example/paper_example.jl
@@ -1,4 +1,4 @@
-using Distributions, LinearAlgebra, Random, Plots
+using Distributions, LinearAlgebra, Random
 using ExponentialFamily, ExponentialFamilyManifolds
 using Manopt, ForwardDiff, StableRNGs
 using PGFPlotsX, RecursiveArrayTools
@@ -87,50 +87,6 @@ function create_target_distribution()
 
     return mixture
 end
-
-"""
-Helper function to visualize and compare two different approximations
-to the target distribution.
-"""
-function plot_comparison(q1, q2, target_dist, title_q1, title_q2)
-    x = range(-8, 8; length=100)
-    y = range(-8, 8; length=100)
-
-    z_target = [pdf(target_dist, [xi, yi]) for yi in y, xi in x]
-    z_p = [pdf(q1, [xi, yi]) for yi in y, xi in x]
-    z_q = [pdf(q2, [xi, yi]) for yi in y, xi in x]
-
-    plt1 = contour(
-        x,
-        y,
-        z_target;
-        fill=true,
-        alpha=0.4,
-        color=:viridis,
-        levels=15,
-        title=title_q1,
-        xlabel="x1",
-        ylabel="x2",
-    )
-    contour!(plt1, x, y, z_p; color=:red, fill=false, linewidth=2, levels=10)
-
-    plt2 = contour(
-        x,
-        y,
-        z_target;
-        fill=true,
-        alpha=0.4,
-        color=:viridis,
-        levels=15,
-        title=title_q2,
-        xlabel="x1",
-        ylabel="x2",
-    )
-    contour!(plt2, x, y, z_q; color=:blue, fill=false, linewidth=2, levels=10)
-
-    return plot(plt1, plt2; layout=(1, 2), size=(1200, 400))
-end
-
 # Example usage reproducing paper figures
 target_dist = create_target_distribution()
 samples = rand(StableRNG(42), target_dist, 2000);
@@ -143,8 +99,6 @@ M = ExponentialFamilyManifolds.get_natural_manifold(form, (2,))
 
 q_mle = optimize_multiple_starts(M, (M, p) -> f_mle(M, p, samples), 2);
 q_vi = optimize_multiple_starts(M, (M, p) -> f_vi(M, p, target_dist), 1);
-plot_comparison(q_mle, q_vi, target_dist, "Maximum likelihood", "Free Energy")
-savefig("comparison1.pdf")
 
 # Isotropic vs Full covariance (Figure 1 bottom)
 p_init_diag = ArrayPartition([0.4, 0.4], [0.5])
@@ -153,5 +107,9 @@ p_isotropic = optimize(M_diag, (M, η) -> f_vi(M, η, target_dist), p_init_diag)
 q_isotropic = convert(
     Distribution, convert(ExponentialFamilyDistribution, M_diag, p_isotropic)
 )
+
+include("plot_helper.jl")
+plot_comparison(q_mle, q_vi, target_dist, "Maximum likelihood", "Free Energy")
+savefig("comparison1.pdf")
 plot_comparison(q_isotropic, q_vi, target_dist, "Isotropic gaussian", "Full covariance")
 savefig("comparison2.pdf")
