### 2025-03-28 c467b40
docs: update paper latex
