+++ paper/paper.tex
@@ -2,87 +2,76 @@
 \setcounter{page}{1}
 
 \usepackage{amsmath}
+\usepackage{footmisc}
 \usepackage{etoolbox}
 \BeforeBeginEnvironment{lstlisting}{\vspace{-\baselineskip}}
 \AfterEndEnvironment{lstlisting}{\vspace{-\baselineskip}}
 
 % \usepackage{todonotes}
 % \usepackage[disable]{todonotes}
-% \newcommand{\ml}[2][] {\todo[inline,backgroundcolor=orange!20!white, #1]{(ml) #2}}
+
 % \newcommand{\bdv}[2][] {\todo[inline,backgroundcolor=blue!20!white, #1]{(bdv) #2}}
 % \newcommand{\ismail}[2][] {\todo[inline,backgroundcolor=red!20!white, #1]{(ismail) #2}}
 % \newcommand{\dmitry}[2][] {\todo[inline,backgroundcolor=teal!20!white, #1]{(dmitry) #2}}
-
-\newcommand{\code}[1]{\textsf{#1}}
 
 \begin{document}
 \input{header}
 
 \maketitle
 \begin{abstract}
-\texttt{ExponentialFamilyManifolds} implements exponential family natural parameter spaces as Riemannian manifolds, enabling geometric optimization over probability distributions. The package automatically manages parameter constraints, such as ensuring the positive definiteness of normal distribution covariance matrices, while providing advanced optimization techniques by integrating \texttt{ExponentialFamily} with \texttt{Manifolds}. Applications for maximum likelihood estimation and variational inference demonstrate the package's practical utility.
+\texttt{ExponentialFamilyManifolds.jl} implements exponential family natural parameter spaces as Riemannian manifolds, enabling geometric optimization over probability distributions. The package automatically manages parameter constraints, such as ensuring the positive definiteness of normal distribution covariance matrices. By representing exponential family distributions as manifolds that conform to the \texttt{ManifoldsBase.jl} interface, it allows users to leverage optimization techniques from \texttt{Manopt} for these manifolds. Applications in maximum likelihood estimation and variational inference highlight the package's practical utility.
+% while providing advanced optimization techniques by integrating \texttt{ExponentialFamily} with \texttt{Manifolds}. Applications for maximum likelihood estimation and variational inference demonstrate the package's practical utility.
 \end{abstract}
 
 \section{Introduction}
 
-Probabilistic inference problems can often be formulated as optimization problems, such as maximum likelihood estimation (MLE) or variational inference (VI) \cite{blei_variational_2017}. In practice, this often involves optimizing over a space of probability distributions to find one that best fits the data (MLE) or approximates a target distribution (VI). Unfortunately, direct optimization over arbitrary probability distributions is intractable. To address these challenges, we often resort to solutions within parametric families, with the exponential family of distributions being a popular choice. This family includes many common distributions, such as Gaussian, Dirichlet, and Wishart distributions. Any member of this family can be written as
[…]
