+++ paper/ref.bib
@@ -10,14 +10,13 @@
   publisher={SIAM}
 }
 
-@software{Senoz_ExponentialFamily_jl_2023,
-author = {Şenöz, İsmail and Bagaev, Dmitry and Lukashchuk, Mykola},
-license = {MIT},
-month = oct,
-title = {{ExponentialFamily.jl}},
-url = {https://github.com/ReactiveBayes/ExponentialFamily.jl},
-version = {1.6.0},
-year = {2023}
+@article{Senoz_ExponentialFamily_jl_2023,
+  author  = {Şenöz, İsmail and Bagaev, Dmitry and Lukashchuk, Mykola},
+  title   = {{ExponentialFamily.jl}},
+  journal = {GitHub repository},
+  year    = {2023},
+  url     = {https://github.com/ReactiveBayes/ExponentialFamily.jl},
+  doi = {10.5281/zenodo.15101588}
 }
 
 @article{axen_manifoldsjl_2023,
@@ -27,7 +26,6 @@ year = {2023}
 	number = {4},
 	journal = {ACM Transactions on Mathematical Software},
 	author = {Axen, Seth D. and Baran, Mateusz and Bergmann, Ronny and Rzecki, Krzysztof},
-	month = dec,
 	year = {2023},
 }
 
@@ -40,7 +38,6 @@ year = {2023}
 	journal = {Journal of Open Source Software},
 	author = {Bergmann, Ronny},
 	year = {2022},
-	note = {Publisher: The Open Journal},
 	pages = {3866},
 }
 
@@ -51,9 +48,7 @@ year = {2023}
 	urldate = {2018-07-22},
 	journal = {arXiv:1607.07892 [cs]},
 	author = {Revels, Jarrett and Lubin, Miles and Papamarkou, Theodore},
-	month = jul,
 	year = {2016},
-	note = {arXiv: 1607.07892},
 	keywords = {Computer Science - Mathematical Software},
 }
 
@@ -67,7 +62,6 @@ year = {2023}
 	number = {2},
 	journal = {Neural Computation},
 	author = {Amari, Shun-ichi},
-	month = jan,
 	year = {1998},
 	pages = {251--276},
 }
@@ -84,10 +78,23 @@ year = {2023}
 	urldate = {2024-11-13},
 	journal = {Journal of the American Statistical Association},
 	author = {Blei, David M. and Kucukelbir, Alp and McAuliffe, Jon D.},
-	month = apr,
 	year = {2017},
-	note = {Publisher: ASA Website
-\_eprint: https://doi.org/10.1080/01621459.2017.1285773},
 	keywords = {Algorithms, Computationally intensive methods, Statistical computing},
 	pages = {859--877},
+}
+
+
+@article{wainwright_graphical_2008,
+	title = {Graphical Models, Exponential Families, and Variational Inference},
+	volume = {1},
+	issn = {1935-8237, 1935-8245},
+	url = {https://www.nowpublishers.com/article/Details/MAL-001},
+	doi = {10.1561/2200000001},
+	abstract = {Graphical Models, Exponential Families, and Variational Inference},
+	pages = {1--305},
+	number = {1},
+	journal = {Foundations and Trends in Machine Learning},
+	author = {Wainwright, Martin J. and Jordan, Michael I.},
+	urldate = {2018-11-26},
+	year = {2008},
 }
