### 2025-03-28 62a9df8
fix: remove standalone function it's fine in the paper
