### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @Nimrais
@kellertuer Thanks for the review. I missed when you added the final part - thanks!

1. **Package Names**: I don't have any strong opinion about it, so I just updated the paper as you suggested.

2. **Advanced Optimization Techniques**: You are correct, I rewrote the abstract to better match this description.

3. **Exponential Family Reference**: Yeah, it was a bit self-obvious for me. But I have added a reference to the already classical book.

4. **Natural Parameters**: I think the paragraph where it's introduced now actually answers your question. I rewrote the sentence. Yes, it's a known term in statistics - another popular choice is to say "canonical parameters".

5. **ManifoldsBase.jl Interface**: Done, I wrote a note that the package implements the interface for manifolds provided in ManifoldsBase.jl.

6. **Type Hierarchy**: I think more precisely it would be `ManifoldsBase.AbstractDecoratorManifold`.

7. **ProductManifold Source**: Done, I mention now that it’s from `ManifoldsBase.jl`

8. **get_natural_manifold method**: I think because it's really important I have explained it more carefully in the paper now.

9. **Retraction Method**: I think I managed to word it better now. The idea of the package is to use optimization from the base manifold - either implemented by me using some numerical geodesic or inherited from the concrete manifold implementation.

10. **Package Dependencies**: I think I found a right way to mention both ManifoldsBase and Manifolds.

11. **Gradient Computation**: I think you are completely correct here. What saved me here is that natural parameters for exponential family in minimal form are convex sets, and it just produced slightly wrong gradients. So I did update this in the paper and in the example code.

12. **Statistical Background**: Yeah, I expect the reader to know what is Variational Inference and Maximum Likelihood estimation already.

13. **References**: Done, thanks!

Stylistic issues (and one grammar from your attached pdf) are also addressed.

Thank you again for your detailed review.

I attach the latest pdf from paper branch here.

<!-- Uploading "paper.pdf"... —>
