### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @Nimrais
Take note that the text has been updated accordingly to the suggestions from another reviewer, and it actually talks about the code in this [PR](https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl/pull/35).
