### 2025-03-28 df055cc
fix: update README with getting started
