### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @Nimrais
And for the code part

14. Yes, we have significantly improved our README documentation at https://github.com/ReactiveBayes/ExponentialFamilyManifolds.jl/blob/paper/README.md. The README now includes:
    - Clear installation instructions
    - A comprehensive getting started guide
    - Basic usage examples
    - Links to the documentation

15. Thank you for confirming!

16-17. Thank you for pointing out these issues. We have addressed them as discussed in point 11.

18. Regarding the supported types suggestion: While we could explicitly list all supported types, our package is designed to work with any distribution type that is supported by `ExponentialFamily.jl`. However, we will consider adding a section in the documentation that lists currently tested distribution types for reference.

19. Yes, we have addressed this in the current README.
