### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @kellertuer
Looks good overall. Just three small comments

* In the abstract `Manopt` should probably read `Manopt.jl`?
* in the keywords “Manifold Optimization” might be misleading for some, since you are not optimising the manifold itself (maybe in the sense of fitting its surface to something) – it is more like “Optimization on Manifolds”?
* Section 2.2 directly before the first code block “along the gradient direction” is at least misleading of not wrong. It is the _negative_ gradient direction!

I do agree with Oliviér's remarks and it is very good that you removed the overhead of additional manifolds that were not really necessary.
