+++ paper/paper.tex
@@ -19,7 +19,7 @@
 
 \maketitle
 \begin{abstract}
-\texttt{ExponentialFamilyManifolds.jl} implements exponential family natural parameter spaces as Riemannian manifolds, enabling geometric optimization over probability distributions. The package automatically manages parameter constraints, such as ensuring the positive definiteness of precision matrices for normal distributions. By representing exponential family distributions as manifolds that conform to the \texttt{ManifoldsBase.jl} interface, it allows users to leverage optimization techniques from \texttt{Manopt} for these manifolds. Applications in maximum likelihood estimation and variational inference highlight the package's practical utility.
+\texttt{ExponentialFamilyManifolds.jl} implements exponential family natural parameter spaces as Riemannian manifolds, enabling geometric optimization over probability distributions. The package automatically manages parameter constraints, such as ensuring the positive definiteness of precision matrices for normal distributions. By representing exponential family distributions as manifolds that conform to the \texttt{ManifoldsBase.jl} interface, it allows users to leverage optimization techniques from \texttt{Manopt.jl} for these manifolds. Applications in maximum likelihood estimation and variational inference highlight the package's practical utility.
 % while providing advanced optimization techniques by integrating \texttt{ExponentialFamily} with \texttt{Manifolds}. Applications for maximum likelihood estimation and variational inference demonstrate the package's practical utility.
 \end{abstract}
 
@@ -29,7 +29,7 @@ Probabilistic inference problems can often be formulated as optimization problem
 \begin{equation*}\label{eq:exponential-family}
 p(x|\eta) = h(x)\exp\left(\eta^\top T(x) - A(\eta)\right),
 \end{equation*}
-where $\eta$ is called natural parameter. Although this parametrization makes optimization more tractable, challenges remain due to the geometric constraints on the parameter space. For example, the natural parameter vector $\eta$ for a multivariate Normal distribution includes a matrix component proportional to the negative precision matrix.
+where $\eta$ is called the natural parameter. Although this parametrization makes optimization more tractable, challenges remain due to the geometric constraints on the parameter space. For example, the natural parameter vector $\eta$ for a multivariate Normal distribution includes a matrix component proportional to the negative precision matrix.
 
 We present \texttt{ExponentialFamilyManifolds.jl}, a Julia \cite{bezanson2017julia} package that addresses these challenges by implementing the parameter spaces as Riemannian manifolds. By bridging \texttt{ExponentialFamily.jl} \cite{Senoz_ExponentialFamily_jl_2023} with the manifold interface from \texttt{ManifoldsBase.jl} \cite{axen_manifoldsjl_2023} and concrete manifolds from \texttt{Manifolds.jl}, the new package enables geometrically aware optimization techniques to handle parameter constraints automatically using \texttt{Manopt.jl} \cite{bergmann_manoptjl_2022}.
 
@@ -40,13 +40,13 @@ In this section, we introduce the \texttt{NaturalParametersManifold} interface, 
 
 A key feature of the core interface \texttt{NaturalParametersManifold} from \texttt{ExponentialFamilyManifolds.jl} is that any point on a manifold—used within \texttt{Manopt.jl}—can also be viewed as an exponential-family distribution. This dual view means we can invoke typical distribution operations (e.g., \texttt{rand} or \texttt{logpdf}) on the same object that supports geometric operations like \texttt{retract}. In other words, we can both \emph{manipulate} a point as part of a manifold-based optimization routine, and use it to \emph{evaluate} functions that treat a point as a distribution for operations like likelihood and sampling.
 
[…]
