### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @Nimrais
Thank you for your careful review and helpful feedback once again.

1. Yes, you are definitely right.
2. Thanks for this clarification. I had thought these two phrases were synonymous, as I've seen several papers use "Manifold Optimization" when they mean "Optimization on Manifolds." Your suggested wording is definitely better and less ambiguous.
3. Ah... Thanks for the correction!

Yes, he definitely made a good point. It is at least far easier to support any breaking changes from `Manifolds.jl`.

[paper.pdf](https://github.com/user-attachments/files/19511532/paper.pdf)
