### 2025-03-28 ReactiveBayes/ExponentialFamilyManifolds.jl#27 @kellertuer
Nice. We do hope to bot have many Breaking changes. For example the next one (about groups) will not affect you.

But sure - if - the new system is easier for you to maintain.
