+++ test/natural_manifolds/wishart_tests.jl
@@ -16,7 +16,10 @@ end
     using DifferentiationInterface
     using FiniteDifferences
 
-    test_mle_works(; mle_samples=500, ndistributions=1, backend_type=AutoFiniteDifferences(central_fdm(5, 1)), kl_friendly=false) do _
-        return WishartFast(4, diagm([1, 1]))
+    test_mle_works(; mle_samples=500, ndistributions=2, backend_type=AutoFiniteDifferences(central_fdm(5, 1)), kl_friendly=false) do rng
+        k = 2
+        L = LowerTriangular(randn(rng, k, k))
+        C = L * L' + k * I
+        return WishartFast(3 + rand(rng), C)
     end
 end
