### 2025-04-02 ReactiveBayes/ExponentialFamilyManifolds.jl#28 @Nimrais
@olivierverdier We need to move forward with updates in other repos. Also, because this release is breaking, we would need to adapt all future PRs. So I suggest that once you are ready, close the issue, not the PR.
