+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"","page":"Home","title":"Home","text":"CurrentModule = ExponentialFamilyManifolds","category":"page"},{"location":"#ExponentialFamilyManifolds","page":"Home","title":"ExponentialFamilyManifolds","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"ExponentialFamilyManifolds.jl provides implementations of manifolds for the natural parameters of exponential family distributions, using Manifolds.jl. These manifolds are compatible with ManifoldsBase.jl, enabling optimization of the natural parameters of exponential family distributions using Manopt.jl.","category":"page"},{"location":"","page":"Home","title":"Home","text":"The primary operation in the package is the get_natural_manifold function, which returns the appropriate manifold for the natural parameters of a given exponential family member type, its dimension and (if required) conditioner. ","category":"page"},{"location":"","page":"Home","title":"Home","text":"ExponentialFamilyManifolds.get_natural_manifold","category":"page"},{"location":"#ExponentialFamilyManifolds.get_natural_manifold","page":"Home","title":"ExponentialFamilyManifolds.get_natural_manifold","text":"get_natural_manifold(::Type{T}, dims, conditioner = nothing)\n\nThe function returns a corresponding manifold for the natural parameters of distribution of type T. Optionally accepts the conditioner, which is set to nothing by default. Use empty tuple () for univariate distributions. \n\njulia> using ExponentialFamily, ExponentialFamilyManifolds\n\njulia> ExponentialFamilyManifolds.get_natural_manifold(Beta, ()) isa ExponentialFamilyManifolds.NaturalParametersManifold\ntrue\n\njulia> ExponentialFamilyManifolds.get_natural_manifold(MvNormalMeanCovariance, (3, )) isa ExponentialFamilyManifolds.NaturalParametersManifold\ntrue\n\n\n\n\n\n","category":"function"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"","page":"Home","title":"Home","text":"The get_natural_manifold function returns NaturalParametersManifold manifold, which is a wrapper around the actual manifold (the base) for the natural parameters that stores extra useful properties and provides the necessary operations for optimization with Manopt.jl. ","category":"page"},{"location":"","page":"Home","title":"Home","text":"using ExponentialFamilyManifolds, ExponentialFamily\n\nExponentialFamilyManifolds.get_natural_manifold(Beta, (), nothing)","category":"page"},{"location":"","page":"Home","title":"Home","text":"ExponentialFamilyManifolds.NaturalParametersManifold","category":"page"},{"location":"#ExponentialFamilyManifolds.NaturalParametersManifold","page":"Home","title":"ExponentialFamilyManifolds.NaturalParametersManifold","text":"NaturalParametersManifold(::Type{T}, dims, base, conditioner)\n\nThe manifold for the natural parameters of the distribution of type T with dimensions dims. An internal structure, use get_natural_manifold to create an instance of a manifold for the natural parameters of distribution of type T.\n\n\n\n\n\n","category":"type"},{"location":"","page":"Home","title":"Home","text":"Its not advised to use the NaturalParametersManifold to create a manifold, but instead use the get_natural_manifold function.","category":"page"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"#Natural-man
[…]
