### 2025-04-11 #179 @lucaferranti
@Nimrais congrats on your work, we are very close! To finalize the publication, please check the checklist in the message above, in a nutshell you should:

- double check authors and affiliations
- make a new release for the version containing the review comments and archive it on zenodo.
- ping me here with the version and zenodo DOI.

Regarding the [references check](https://github.com/JuliaCon/proceedings-review/issues/179#issuecomment-2796473960), there seems to be one reference without DOI. That's ok if there really isn't one available, but I think that paper [has one](https://arxiv.org/abs/1607.07892), can you check?

Once you've fixed the references and published the new release, ping me and I'll finalize the acceptance
