### 2025-04-11 761baf0
fix: make zenodo happy with an ad-hoc release
