### 2025-04-11 #179 @Nimrais
@lucaferranti Thanks!

The authors and affiliations are correct, and I have created a version after the review https://zenodo.org/records/15196176. The Zenodo DOI is https://doi.org/10.5281/zenodo.15196175.

Regarding the ForwardDiff paper, I've verified that it appears to be only available as an arXiv preprint without a formal DOI. The citation format provided in their repository only includes the arXiv identifier
```bibtex
@article{RevelsLubinPapamarkou2016,
    title = {Forward-Mode Automatic Differentiation in {J}ulia},
   author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
  journal = {arXiv:1607.07892 [cs.MS]},
     year = {2016},
      url = {https://arxiv.org/abs/1607.07892}
}
```

I checked several papers that cite this work, for example, this one https://www.jmlr.org/papers/volume18/17-468/17-468.pdf, and they also cite it without a DOI. The current citation format is appropriate unless you have information suggesting it was later published in a journal with a DOI. I also searched Zendo for it but couldn't find it there.
