### 2025-04-11 #179 @lucaferranti
thanks @Nimrais can you confirm the correct version is 3.0.1?
