### 2025-04-11 #179 @Nimrais
I am not sure. It's 3.0.0 (the PR that resolved the review was there), but in Zenodo, there is no way to backtrack to versions we created before syncing the repo with Zenodo. So, to create a Zenodo tracking release, I did an ad-hoc fix release. It's different from 3.0.0 in only one line, so from the point of reproducibility of the paper, it's the same.
