### 2025-04-11 #179 @Nimrais
So, from my point of view, we can safely select 3.0.1.
